Added a chronos clock that works

+++ MortalRescue/src/GameObject.h
@@ -16,7 +16,7 @@ public:
 	void update();
 	void init();
 
-	int xPos, yPos, xSize, ySize, speed, xVelocity, yVelocity;
+	float xPos, yPos, xSize, ySize, speed, xVelocity, yVelocity;
 	string id, description;
 	bool isStaticObject, isAnimated;
 	SDL_Texture* staticTexture; //Used if this is not an animated object
